+++ crates/api-snowflake-rest/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod auth;
 pub mod error;
 pub mod handlers;
+pub mod layer;
 pub mod router;
 pub mod schemas;
 pub mod state;
